modify bug of pilicy tree and data collection function

+++ backend/apolo/apolomgr/resource/collection_policy_tree/policytree_view.py
@@ -15,7 +15,7 @@ from django.db import transaction
 from rest_framework import viewsets
 
 from backend.apolo.apolomgr.resource.common.common_policy_tree.policy_tree import Policy_tree
-from backend.apolo.models import CollPolicy, CollPolicyRuleTree, Items
+from backend.apolo.models import CollPolicy, CollPolicyRuleTree, Items, PolicysGroups
 from backend.apolo.serializer.policytree_serializer import CollPolicyRuleTreeSerializer
 from backend.apolo.tools import views_helper, constants
 from backend.apolo.tools.exception import exception_handler
@@ -96,39 +96,58 @@ class PolicyTreeViewSet(viewsets.ViewSet):
         self.tree = views_helper.get_request_value(self.request, 'tree', 'BODY')
         self.raw_data = views_helper.get_request_value(self.request, 'raw_data', 'BODY')
         try:
-
-            tree_is_exits = CollPolicyRuleTree.objects.filter(coll_policy=self.coll_policy_id)
-            with transaction.atomic():
-                if tree_is_exits:
-                    # check the policy is exits in the item tables
-                    items_list = Items.objects.filter(coll_policy=self.coll_policy_id)
-                    if len(items_list) > 0:
-                        data = {
-                            'data': '',
-                            'new_token': self.new_token,
-                            constants.STATUS: {
-                                constants.STATUS: constants.FALSE,
-                                constants.MESSAGE: constants.POLICY_IS_APPLIED
-                            }
-                        }
-                        return api_return(data=data)
-                    else:
-                        # del the old policy tree
-                        CollPolicyRuleTree.objects.filter(coll_policy=self.coll_policy_id).delete()
-
-
-                # save the new policy
-                # update cli_command_result into coll_policy table
-                coll_policy = CollPolicy.objects.get(pk=self.coll_policy_id)
-                if self.raw_data:
-                    coll_policy.cli_command_result = self.raw_data
-                    coll_policy.save()
-                    # select all nodes of the policy tree
-                policy_tree = Policy_tree(self.coll_policy_id)
-                policy_tree.get_all_nodes(self.tree)
-                obj = policy_tree.all_nodes
-                data = self.__save_the_policy_tree(nodes_dict=obj)
+            if not self.__check_policy_tree_in_group():
+                data = {
+                    'new_token': self.new_token,
+                    constants.STATUS: {
+                        constants.STATUS: constants.FALSE,
+                        constants.MESSAGE: constants.POLICY_TREE_IS_GROUPED
+                    }
+                }
                 return api_return(data=data)
+            else:
+                if self.__check_the_node_is_leaf(self.tree):
+
+                    tree_is_exits = CollPolicyRuleTree.objects.filter(coll_policy=self.coll_policy_id)
+                    with transaction.atomic():
+                        if tree_is_exits:
+                            # check the policy is exits in the item tables
+                            items_list = Items.objects.filter(coll_policy=self.coll_policy_id)
+                            if len(items_list) > 0:
+                                data = {
+                                    'data': '',
+                                    'new_token': self.new_token,
+                                    constants.STATUS: {
+                                        constants.STATUS: constants.FALSE,
+                                        constants.MESSAGE: constants.POLICY_IS_APPLIED
+                                    }
+                                }
+                                return api_return(data=data)
+                            else:
+                                # del the old policy tree
+                                CollPolicyRuleTree.objects.filter(coll_policy=self.coll_policy_id).delete()
+
+                        # save the new policy
+                        # update cli_command_result into coll_policy table
+                        coll_policy = CollPolicy.objects.get(pk=self.coll_policy_id)
+                        if self.raw_data:
+                            coll_policy.cli_command_result = self.raw_data
+                            coll_policy.save()
+                            # select all nodes of the policy tree
+                        policy_tree = Policy_tree(self.coll_policy_id)
+                        policy_tree.get_all_nodes(self.tree)
+                        obj = policy_tree.all_nodes
+                        data = self.__save_the_policy_tree(nodes_dict=obj)
+                        return api_return(data=data)
+                else:
+                    data = {
+                        'new_token': self.new_token,
+                        constants.STATUS: {
+                            constants.STATUS: constants.FALSE,
+                            constants.MESSAGE: constants.LEAF_IS_BLOCK_RULE
+                        }
+                    }
+                    return api_return(data=data)
         except Exception as e:
             print e
             data = {
@@ -193,19 +212,25 @@ class PolicyTreeViewSet(viewsets.ViewSet):
         return data
 
     def __check_the_node_is_leaf(self, tree_dict):
-
-        rule_type = 0
-        if tree_dict['data']['rule_type']:
-           rule_type = int(tree_dict['data']['rule_type'].split('_')[2])
+        rule_type = tree_dict['data']['rule_type'].split('_')[0]
+        print rule_type
         if tree_dict.has_key('children'):
             if len(tree_dict['children']) > 0:
+                block_is_leaf = False
                 for children in tree_dict['children']:
-                    self.__check_the_node_is_leaf(children)
+                     block_is_leaf = self.__check_the_node_is_leaf(children)
+                     if not block_is_leaf:
+                         break
+                return block_is_leaf
             else:
-                if 0 < rule_type < 5:
-                     return True
+                if rule_type =='block':
+                    return False
                 else:
-                     return False
+                    return True
 
-
-
+    def __check_policy_tree_in_group(self):
+        queryset = PolicysGroups.objects.filter(policy=self.coll_policy_id)
+        if queryset.count()> 0:
+            return False
+        else:
+            return True
